backend/requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.11.0
attrs==25.1.0
bitarray==3.0.0
certifi==2025.1.31
charset-normalizer==3.4.1
ckzg==2.0.1
click==8.1.8
cytoolz==1.0.1
distro==1.9.0
eth-account==0.13.5
eth-hash==0.7.1
eth-keyfile==0.8.1
eth-keys==0.6.1
eth-rlp==2.2.0
eth-typing==5.1.0
eth-utils==5.2.0
eth_abi==5.2.0
fastapi==0.115.8
frozenlist==1.5.0
h11==0.14.0
hexbytes==1.3.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
multidict==6.1.0
openai==1.61.1
parsimonious==0.10.0
propcache==0.2.1
pycryptodome==3.21.0
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
pyunormalize==16.0.0
regex==2024.11.6
requests==2.32.3
rlp==4.1.0
scheduler==0.8.8
sniffio==1.3.1
starlette==0.45.3
toolz==1.0.0
tqdm==4.67.1
typeguard==4.4.1
types-requests==2.32.0.20241016
typing_extensions==4.12.2
tzlocal==5.2
urllib3==2.3.0
uvicorn==0.34.0
web3==7.8.0
websockets==13.1
yarl==1.18.3
